fix: revert to original tracer config when stops
