+++ crates/cheatcodes/src/inspector.rs
@@ -36,7 +36,7 @@ use foundry_evm_core::{
     utils::new_evm_with_existing_context,
     InspectorExt,
 };
-use foundry_evm_traces::TracingInspector;
+use foundry_evm_traces::{TracingInspector, TracingInspectorConfig};
 use itertools::Itertools;
 use proptest::test_runner::{RngAlgorithm, TestRng, TestRunner};
 use rand::Rng;
@@ -75,12 +75,6 @@ pub trait CheatcodesExecutor {
         &'a mut self,
         cheats: &'a mut Cheatcodes,
     ) -> impl InspectorExt<DB> + 'a;
-
-    fn start_steps_recording(&mut self, _cheats: &mut Cheatcodes) {}
-
-    fn stop_and_get_recorded_step(&mut self, _cheats: &mut Cheatcodes) -> Vec<&CallTraceStep> {
-        Vec::new()
-    }
 
     /// Constructs [revm::Evm] and runs a given closure with it.
     fn with_evm<DB: DatabaseExt, F, O>(
@@ -227,6 +221,14 @@ pub struct BroadcastableTransaction {
     pub rpc: Option<String>,
     /// The transaction to broadcast.
     pub transaction: TransactionMaybeSigned,
+}
+
+#[derive(Clone, Debug, Copy)]
+pub struct RecordDebugStepInfo {
+    /// The debug trace node index when the recording starts.
+    pub start_node_idx: usize,
+    /// The original tracer config when the recording starts.
+    pub original_tracer_config: TracingInspectorConfig,
 }
 
 /// Holds gas metering state.
@@ -404,8 +406,8 @@ pub struct Cheatcodes {
     /// It is empty if nothing were recorded.
     pub recorded_debug_steps: Option<Vec<crate::Vm::DebugStep>>,
 
-    /// The tracing node index where we start to record the debug trace steps.
-    pub record_debug_steps_start_index: Option<usize>,
+    /// The information of the debug step recording.
+    pub record_debug_steps_info: Option<RecordDebugStepInfo>,
 
     /// Recorded logs
     pub recorded_logs: Option<Vec<crate::Vm::Log>>,
@@ -500,7 +502,7 @@ impl Cheatcodes {
             recorded_account_diffs_stack: Default::default(),
             recorded_logs: Default::default(),
             recorded_debug_steps: Default::default(),
-            record_debug_steps_start_index: Default::default(),
+            record_debug_steps_info: Default::default(),
             mocked_calls: Default::default(),
             mocked_functions: Default::default(),
             expected_calls: Default::default(),
